refact: URL検出をインフラ層で行う
