+++ workspaces/server/src/repository/social-link.ts
@@ -1,7 +1,7 @@
 export interface ISocialLink {
 	userId: string;
 	links: {
-		providerId: number;
+		providerId: string;
 		handle: string;
 		url: string;
 	}[];
@@ -11,5 +11,5 @@ export interface ISocialLinkRepository {
 	getSocialLinksByUserId: (
 		userId: string,
 	) => Promise<ISocialLink>;
-	updateSocialLinks: (socialLinks: ISocialLink) => Promise<void>;
+	updateSocialLinks: (userId: string, links: string[]) => Promise<void>;
 }
